Refactor namespace structure for DialogFlow.Sdk
