+++ DialogFlow.Sdk/Builders/ParameterBuilder.cs
@@ -1,4 +1,4 @@
-﻿using DialogFlow.Sdk.Intents;
+﻿using DialogFlow.Sdk.Models.Intents;
 
 namespace DialogFlow.Sdk.Builders
 {
