+++ DialogFlow.Sdk/Builders/TriggerStatementParser.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using DialogFlow.Sdk.Intents;
+using DialogFlow.Sdk.Models.Intents;
 
 namespace DialogFlow.Sdk.Builders
 {
