+++ DialogFlow.Sdk/DialogFlow.Sdk.csproj
@@ -6,4 +6,7 @@
   <ItemGroup>
     <PackageReference Include="Refit" Version="4.3.0" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Api" />
+  </ItemGroup>
 </Project>
