+++ DialogFlow.Sdk/DialogFlowClient.cs
@@ -1,7 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using DialogFlow.Sdk.Intents;
+using DialogFlow.Sdk.Models.Intents;
+using DialogFlow.Sdk.Rest;
 using Refit;
 
 namespace DialogFlow.Sdk
