+++ DialogFlow.Sdk/Models/Fulfillment/FulfillmentRequest.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using Newtonsoft.Json;
 
-namespace DialogFlow.Sdk.Fulfillment
+namespace DialogFlow.Sdk.Models.Fulfillment
 {
     public class FulfillmentRequest
     {
