+++ DialogFlow.Sdk/Models/Intents/Events.cs
@@ -1,4 +1,4 @@
-﻿namespace DialogFlow.Sdk.Intents
+﻿namespace DialogFlow.Sdk.Models.Intents
 {
     public class Events
     {
