+++ DialogFlow.Sdk/Rest/IDialogFlowApi.cs
@@ -1,9 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using DialogFlow.Sdk.Intents;
+using DialogFlow.Sdk.Models.Intents;
 using Refit;
 
-namespace DialogFlow.Sdk
+namespace DialogFlow.Sdk.Rest
 {
     public interface IDialogFlowApi
     {
