+++ Jobber.SmartAssistant/Features/CreateJob/ClientRequestedCreateJobIntentFulfiller.cs
@@ -1,8 +1,7 @@
 ﻿using System.Linq;
 using System.Threading.Tasks;
 using DialogFlow.Sdk.Builders;
-using DialogFlow.Sdk.Fulfillment;
-using DialogFlow.Sdk.Intents;
+using DialogFlow.Sdk.Models.Fulfillment;
 using Jobber.Sdk;
 using Jobber.Sdk.Models;
 using Jobber.Sdk.Responses;
