+++ Jobber.SmartAssistant/Features/FavoriteNumber/FavoriteNumberIntentDefinition.cs
@@ -1,6 +1,7 @@
 ﻿using Assistant.Sdk.BuiltIns;
 using DialogFlow.Sdk.Builders;
-using DialogFlow.Sdk.Intents;
+using DialogFlow.Sdk.Models.Common;
+using DialogFlow.Sdk.Models.Intents;
 
 namespace Jobber.SmartAssistant.Features.FavoriteNumber
 {
