+++ Jobber.SmartAssistant/Program.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using Assistant.Sdk.BuiltIns;
 using Assistant.Sdk.Core;
 using DialogFlow.Sdk;
+using DialogFlow.Sdk.Rest;
 using Jobber.SmartAssistant.Core;
 using Jobber.SmartAssistant.Features.CreateJob;
 using Jobber.SmartAssistant.Features.Fallback;
